// src/store/filter/types.ts
export enum SortTypeEnum {
  RATING_DESC = 'rating',
  RATING_ASC = '-rating',
  TITLE_DESC = 'title',
  TITLE_ASC = '-title',
  PRICE_DESC = 'price ',
  PRICE_ASC = '-price ',
}

export type SortProperty = {
  name: string
  sortType: SortTypeEnum
}

export interface FilterSliceState {
  searchValue: string
  categoryId: number,
  currentPage: number,
  sort: SortProperty
}